Ask for confirmation only after the resource has been selected

+++ commands/core/delete.sh
@@ -3,24 +3,30 @@
 function dx::delete_resource() {
   local -r resource_type="${1}"
   local -r resource_id="${2}"
+  local -r description="${3}"
 
-  case "${resource_type}" in
-    'container')
-      docker rm --force "${resource_id}"
-      ;;
-    'image')
-      docker rmi --force "${resource_id}"
-      ;;
-    'volume')
-      docker volume rm --force "${resource_id}"
-      ;;
-    'network')
-      docker network rm "${resource_id}"
-      ;;
-    *)
-      dx::log::errexit "Unsupported resource type (${resource_type})"
-      ;;
-  esac
+  dx::log::info "Do you really want to delete the ${resource_type} [${description} - ${resource_id}]? Be careful. [y/n]"
+  read -e -r confirmation
+
+  if [ "${confirmation}" = 'y' ] || [ "${confirmation}" = 'yes' ]; then
+    case "${resource_type}" in
+      'container')
+        docker rm --force "${resource_id}"
+        ;;
+      'image')
+        docker rmi --force "${resource_id}"
+        ;;
+      'volume')
+        docker volume rm --force "${resource_id}"
+        ;;
+      'network')
+        docker network rm "${resource_id}"
+        ;;
+      *)
+        dx::log::errexit "Unsupported resource type (${resource_type})"
+        ;;
+    esac
+  fi
 }
 
 function dx::delete() {
@@ -33,20 +39,13 @@ function dx::delete() {
       dx::log::errexit 'No resources found'
     fi
 
-    local type=''
-    local id=''
+    local -r row="$(echo -e "${resources}" | fzf)"
 
-    echo -e "${resources}" \
-      | fzf --multi \
-      | while read -r resource; do
+    local -r type="$(echo "${row}" | awk '{ print $1 }')"
+    local -r id="$(echo "${row}" | awk '{ print $2 }')"
+    local -r description="$(echo "${row}" | awk '{ print $3 }')"
 
-        if [ -n "${resource}" ]; then
-          type="$(echo "${resource}" | awk '{ print $1 }')"
-          id="$(echo "${resource}" | awk '{ print $2 }')"
-
-          dx::delete_resource "${type}" "${id}"
-        fi
-      done
+    dx::delete_resource "${type}" "${id}" "${description}"
   else
     export PS3=$'\n''Type the respective resource number: '$'\n'
 
@@ -59,12 +58,14 @@ function dx::delete() {
 
     local type=''
     local id=''
+    local description=''
 
     select resource in "${resources[@]}"; do
       type="$(echo "${resource}" | awk '{ print $1 }')"
       id="$(echo "${resource}" | awk '{ print $2 }')"
+      description="$(echo "${resource}" | awk '{ print $3 }')"
 
-      dx::delete_resource "${type}" "${id}"
+      dx::delete_resource "${type}" "${id}" "${description}"
       break
     done
   fi
